refactor tests code
