+++ app/src/main/java/com/geekbrains/myfirsttests/domain/Car.kt
@@ -1,4 +1,4 @@
-package com.geekbrains.myfirsttests
+package com.geekbrains.myfirsttests.domain
 
 data class Car(
     var nameModel:String? = null,
