+++ app/src/main/java/com/geekbrains/myfirsttests/domain/CarUseCase.kt
@@ -1,4 +1,6 @@
-package com.geekbrains.myfirsttests
+package com.geekbrains.myfirsttests.domain
+
+import com.geekbrains.myfirsttests.domain.Car
 
 interface CarUseCase {
     fun create(car: Car)
